BOCM-358 - moves service code to own file
